BackEnd/requirements.txt
absl-py==1.4.0
asgiref==3.3.4
astunparse==1.6.3
bcrypt==3.2.0
cachetools==5.3.0
certifi==2020.12.5
cffi==1.14.5
charset-normalizer==3.1.0
Django==3.2.18
django-cors-headers==3.7.0
djangorestframework==3.12.4
flatbuffers==23.3.3
gast==0.4.0
google-auth==2.17.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.54.0
h5py==3.8.0
idna==3.4
importlib-metadata==6.6.0
keras==2.11.0
Keras-Preprocessing==1.1.2
libclang==16.0.0
Markdown==3.4.3
MarkupSafe==2.1.2
numpy==1.21.6
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
Pillow==9.5.0
pip==22.3.1
protobuf==3.19.6
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
setuptools==65.6.3
six==1.16.0
sqlparse==0.2.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-io-gcs-filesystem
termcolor==2.3.0
typing-extensions==3.7.4.3
urllib3==1.26.15
Werkzeug==2.2.3
wheel==0.38.4
wrapt==1.15.0
zipp==3.15.0
